Hefty Heimdall 4.0.3 changes.

+++ src/YourMoneroRequests.cpp
@@ -1537,7 +1537,7 @@ YourMoneroRequests::get_tx(
         j_response["mixin_no"]       = mixin_no;
         j_response["num_of_outputs"] = output_pub_keys.size();
         j_response["num_of_inputs"]  = input_key_imgs.size();
-        j_response["tx_version"]     = tx.version;
+        j_response["tx_version"]     = static_cast<uint16_t>(tx.version);
         j_response["rct_type"]       = tx.rct_signatures.type;
 
         if (!coinbase &&  tx.vin.size() > 0)
